+++ src/main/java/io/tarantool/driver/core/proxy/CRUDBaseOptions.java
@@ -1,13 +1,16 @@
 package io.tarantool.driver.core.proxy;
 
+import io.tarantool.driver.core.space.options.Options;
+
 /**
  * This class is not part of the public API.
  *
  * Represent basic options for all cluster operations
  *
  * @author Alexey Kuzin
+ * @author Artyom Dubinin
  */
-public class CRUDBaseOptions extends CRUDAbstractOperationOptions {
+class CRUDBaseOptions extends CRUDAbstractOperationOptions {
 
     public static final String TIMEOUT = "timeout";
 
@@ -16,6 +19,13 @@ public class CRUDBaseOptions extends CRUDAbstractOperationOptions {
     CRUDBaseOptions(AbstractBuilder<O, T> builder) {
         if (builder.timeout != null) {
             addOption(TIMEOUT, builder.timeout);
+        }
+
+        if (builder.options != null) {
+            Object batchSize = builder.options.asMap().get(TIMEOUT);
+            if (batchSize != null) {
+                addOption(TIMEOUT, batchSize);
+            }
         }
     }
 
@@ -28,9 +38,15 @@ public class CRUDBaseOptions extends CRUDAbstractOperationOptions {
     class AbstractBuilder<O extends CRUDBaseOptions, T extends AbstractBuilder<O, T>>
         extends CRUDAbstractOperationOptions.AbstractBuilder<O, T> {
         protected Integer timeout;
+        protected Options options;
 
         public T withTimeout(int timeout) {
             this.timeout = timeout;
+            return self();
+        }
+
+        public T withOptions(Options options) {
+            this.options = options;
             return self();
         }
     }