+++ src/main/java/io/tarantool/driver/core/space/options/proxy/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Contains implementations for Tarantool proxy space options
+ *
+ * @author Alexey Kuzin
+ * @author Artyom Dubinin
+ */
+package io.tarantool.driver.core.space.options.proxy;